+++ includes/vector.h
@@ -3,9 +3,9 @@
 
 typedef struct	s_vec
 {
-	float	x;
-	float 	y;
-	float 	z;
+	double	x;
+	double 	y;
+	double 	z;
 }				t_vec;
 
 typedef	struct	s_ray
@@ -14,13 +14,13 @@ typedef	struct	s_ray
 	t_vec	*dir;
 }				t_ray;
 
-t_vec	*vec_default(float x, float y, float z);
-float	vec_length(t_vec *vec);
+t_vec	*vec_default(double x, double y, double z);
+double	vec_length(t_vec *vec);
 t_vec	*vec_subtract(t_vec *vec1, t_vec *vec2);
 void	vec_normalize(t_vec *vec);
-float	vec_dot_product(t_vec *vec1, t_vec *vec2);
+double	vec_dot_product(t_vec *vec1, t_vec *vec2);
 t_vec	*vec_cross(t_vec *vec1, t_vec *vec2);
 t_vec	*vec_sum(t_vec *vec1, t_vec *vec2);
-t_vec	*vec_multiplication(t_vec *vec, float k);
+t_vec	*vec_multiplication(t_vec *vec, double k);
 
 #endif
